+++ client/hooks/useSearch.ts
@@ -1,6 +1,0 @@
-import qs from "query-string";
-import { useLocation } from "react-router";
-
-export default function useSearch() {
-  return qs.parse(useLocation().search);
-}